+++ tests/common.rs
@@ -1,5 +1,5 @@
 use jubjub::*;
-use rand_core::{RngCore, SeedableRng};
+use rand_core::{Rng, SeedableRng};
 use rand_xorshift::XorShiftRng;
 
 pub const NUM_BLACK_BOX_CHECKS: u32 = 2000;
@@ -9,11 +9,11 @@ pub fn new_rng() -> XorShiftRng {
 }
 
 pub trait MyRandom {
-    fn new_random<T: RngCore>(rng: &mut T) -> Self;
+    fn new_random<T: Rng>(rng: &mut T) -> Self;
 }
 
 impl MyRandom for Fq {
-    fn new_random<T: RngCore>(rng: &mut T) -> Self {
+    fn new_random<T: Rng>(rng: &mut T) -> Self {
         let mut random_bytes = [0u8; 64];
         rng.fill_bytes(&mut random_bytes);
         Fq::from_bytes_wide(&random_bytes)
@@ -21,7 +21,7 @@ impl MyRandom for Fq {
 }
 
 impl MyRandom for Fr {
-    fn new_random<T: RngCore>(rng: &mut T) -> Self {
+    fn new_random<T: Rng>(rng: &mut T) -> Self {
         let mut random_bytes = [0u8; 64];
         rng.fill_bytes(&mut random_bytes);
         Fr::from_bytes_wide(&random_bytes)